+++ packages/plugin-better-auth/server/test/utils.ts
@@ -102,11 +102,13 @@ export async function stopStrapi() {
       "database.connection.connection.filename",
     );
 
-    assert(typeof tmpDbFile === "string");
-
     await instance.destroy();
 
-    if (fs.existsSync(tmpDbFile)) {
+    if (
+      tmpDbFile &&
+      typeof tmpDbFile === "string" &&
+      fs.existsSync(tmpDbFile)
+    ) {
       fs.unlinkSync(tmpDbFile);
     }
